Initial release of Second Generation FPGA-Powered Amazon EC2 Instances (F2)
The AWS EC2 FPGA Development Kit contains the hardware/software tools that you will use to develop, simulate, debug, compile and run your hardware-accelerated FPGA applications.  There is no software charge for the development tools when using the FPGA developer AMI and you can program the FPGAs on your F2 instance as many times as you like with no additional fees.

+++ sdk_setup.sh
@@ -1,5 +1,5 @@
 #
-# Copyright 2015-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may
 # not use this file except in compliance with the License. A copy of the
@@ -22,13 +22,33 @@ if [ $script == $0 ]; then
   exit 2
 fi
 
+export REPO_ROOT=$(pwd)
+
+# Install base packages on any AMI
+UBUNTU_DISTRO='"Ubuntu"'
+DISTRO=$(awk -F= '/^NAME/{print $2}' /etc/os-release)
+echo "Sourcing SDK setup on a '$DISTRO' machine. Next comparing if it is equal to '$UBUNTU_DISTRO'"
+if [[ "$DISTRO" == "$UBUNTU_DISTRO" ]]; then
+  sudo apt update -y
+  sudo apt install -y linux-headers-$(uname -r) bzip2 unzip libx11-6 perl build-essential gdb git screen tmux pciutils libstdc++6 libjpeg-turbo8-dev libtiff5 ocl-icd-opencl-dev opencl-headers tcl tcl-dev wget awscli environment-modules python3-pip
+else
+  sudo yum upgrade -y
+  sudo yum install -y kernel kernel-devel kernel-headers bzip2 unzip libX11 perl gcc gcc-c++ gdb git screen tmux pciutils libstdc++-static libjpeg-turbo-devel libtiff-devel ocl-icd ocl-icd-devel opencl-headers tcl tcl-devel wget awscli environment-modules python3-pip
+fi
+
 full_script=$(readlink -f $script)
 script_name=$(basename $full_script)
 script_dir=$(dirname $full_script)
 current_dir=$(pwd)
 
-source $script_dir/shared/bin/set_common_functions.sh
-source $script_dir/shared/bin/set_common_env_vars.sh
+if ! source $script_dir/shared/bin/set_common_functions.sh; then
+  err_msg "Couldn't set up common functions for SDK! Exiting!"
+  return 1
+fi
+if ! source $script_dir/shared/bin/set_common_env_vars.sh; then
+  err_msg "Couldn't set up env vars for SDK! Exiting!"
+  return 1
+fi
 
 sudo rm -f /tmp/sdk_root_env.exp
 typeset -f allow_non_root > /tmp/sdk_root_env.exp
@@ -51,3 +71,4 @@ fi
 
 cd $current_dir
 info_msg "$script_name PASSED"
+
